bug fix on firm,client,operational cost, and client profile

+++ app/Http/Controllers/Lawyer/ClientAccountController.php
@@ -162,7 +162,7 @@ class ClientAccountController extends Controller
 
         try {
             $clientAccount = ClientAccount::findOrFail($request->id); // Find the record by ID
-            if (!$request->hasFile('upload')) {
+            if ($request->existingDocument != null && $request->upload == null) {
                 if (str_contains("funds in", $request->transaction_type)) {
                     $clientAccount->update([
                         'date' => $request->date,
@@ -191,13 +191,7 @@ class ClientAccountController extends Controller
                 return redirect()->route('lawyer.client-accounts.show', ['client_account' => $request->bank_account_id])->with('successMessage', 'Successfully update the transaction.');
             } else {
                 $fileName = uniqid('TRANSACTION_') . '_' . date('Ymd') . '_' . time() . '.' . $request->file('upload')->extension();
-                $filePath = null;
-
-                if (Storage::exists(ClientAccount::UPLOAD_PATH . $fileName)) {
-                    // DO NOTHING
-                } else {
-                    $filePath = $request->file('upload')->storeAs(ClientAccount::UPLOAD_PATH, $fileName);
-                }
+                $filePath = $request->file('upload')->storeAs(ClientAccount::UPLOAD_PATH, $fileName);
 
                 $request->merge(['upload_filename' => $fileName]);
 
@@ -232,7 +226,7 @@ class ClientAccountController extends Controller
                 return redirect()->route('lawyer.client-accounts.show', ['client_account' => $request->bank_account_id])->with('successMessage', 'Successfully update the transaction.');
             }
         } catch (\Exception $e) {
-            if (Storage::exists($filePath)) {
+            if ($filePath != null && Storage::exists($filePath)) {
                 Storage::delete($filePath);
             }
